Performance: Try to get rid of range comparisons. (the previous commit also removed some trigger points)
It appears that the amount of range queries should be reduced drastically. The only other reason for a frame rate of 1 fps can only be the pathfinder, but since all trees are removed by now, the frame rate is still terrible and thus the we have to somehow profile the range queries.

Generally it appears that the removal of the formations in 0AD has both positive but also negative effects on performance depending on the individual case.